+++ docs/source/api_reference/dagger/config/index.rst
@@ -1,0 +1,7 @@
+.. automodule:: vivarium_cluster_tools.dagger.config
+
+.. toctree::
+   :maxdepth: 2
+   :glob:
+
+   *
