+++ docs/source/api_reference/dagger/config/serialization.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium_cluster_tools.dagger.config.serialization
