--- docs/source/api_reference/dagger/workflow_config/builder.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_cluster_tools.dagger.workflow_config.builder
