--- docs/source/api_reference/dagger/workflow_config/index.rst
@@ -1,7 +1,0 @@
-.. automodule:: vivarium_cluster_tools.dagger.workflow_config
-
-.. toctree::
-   :maxdepth: 2
-   :glob:
-
-   *
