--- docs/source/api_reference/dagger/workflow_config/parsing.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium_cluster_tools.dagger.workflow_config.parsing
