+++ src/vivarium_cluster_tools/dagger/cli.py
@@ -19,7 +19,7 @@ from vivarium.framework.utilities import handle_exceptions
 
 from vivarium_cluster_tools.core import cli_tools, logs
 from vivarium_cluster_tools.dagger import runner
-from vivarium_cluster_tools.dagger.workflow_config.parsing import load_workflow_config
+from vivarium_cluster_tools.dagger.config.parsing import load_workflow_config
 
 
 @click.group()
